chore: move GlobalHandlers
