+++ src/main/java/org/example/delivery/common/GlobalExceptionHandlerDev.java
@@ -1,4 +1,4 @@
-package org.example.delivery.config;
+package org.example.delivery.common;
 
 
 import org.springframework.context.annotation.Profile;
